pyproject.toml
[project]
name = "nodox-curriculum-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
